+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.5.2] - 04.11.2021
+
+* invalid timstamps fixes for data saving
+
+## [1.5.1] - 15.10.2021
+
+* package minor fixes
+
+## [1.5.0] - 15.10.2021
+
+* no stream heartbeatSeriesQuery. HeartbeatSeries now is a valid sample with a set of beat by beat measurements 
+
 ## [1.4.1] - 12.09.2021
 
 * Fix characteristic parsing if not all types are requested for reading